move commands related code from device_driver.c
